Попытка реализации паттерна MVVM к приложению PlaylistMaker

+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackDto.kt
@@ -6,7 +6,7 @@ data class TrackDto(
     val trackId: Int,
     val trackName: String,
     val artistName: String,
-    val trackTimeMillis: Long,
+    val trackTimeMillis: Int,
     val artworkUrl100: String,
     val collectionName: String?,
     val releaseDate: String?,
